+++ CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **BC break:** Rename `Scope::obtain()` to `resolve()`.
 - **BC break:** Rename `DIC::scope()` to `DIC::scopedOf()`.
 - **BC break:** Merge `DIC::bind()` and `bindQualifier()` into single `DIC::bind(mixed $value, Type $type, string|\Stringable|\UnitEnum $qualifier = '')`.
+- **BC break:** Autowiring now matches bindings by exact type only.
+- **BC break:** `DIC::bind()` and `Scoped::with()` now override previous bindings.
 
 ### Removed