+++ CHANGELOG.md
@@ -5,7 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.4.0] - 2026-04-22
+## [0.4.0] - 2026-04-23
+
+### Added
+
+- `DIC::function()` now supports `#[Tag]` attributes on the function.
+- `DIC::function()` now supports `#[Singleton]` and `#[Transient]` lifetime attributes on the function (defaults to `scoped`).
 
 ### Changed